:zap: removing torchapp2
