+++ gambit/apps.py
@@ -1,89 +1,257 @@
-from pathlib import Path
+import psutil
 import torch
-from Bio import SeqIO
+import numpy as np
+from pathlib import Path
 from torch import nn
-from fastai.data.core import DataLoaders
-import torchapp as ta
-from corgi.seqtree import SeqTree, node_to_str
-from rich.console import Console
+import lightning as L
+from corgi.seqtree import SeqTree, SeqDetail
 from seqbank import SeqBank
-
 from hierarchicalsoftmax import HierarchicalSoftmaxLoss, SoftmaxNode
-from hierarchicalsoftmax.metrics import GreedyAccuracy
-from hierarchicalsoftmax.inference import node_probabilities, greedy_predictions, render_probabilities
+from hierarchicalsoftmax.metrics import RankAccuracyTorchMetric
+from torch.utils.data import DataLoader
+from collections.abc import Iterable
+from rich.console import Console
+from torch.utils.data import Dataset
+from Bio import SeqIO
+from dataclasses import dataclass
+from hierarchicalsoftmax.metrics import greedy_accuracy
+from torchmetrics import Metric
 import pandas as pd
+import os
+from hierarchicalsoftmax.inference import node_probabilities, greedy_predictions, render_probabilities
 
+from lightning.pytorch.loggers import CSVLogger
 
+# import torchapp as ta
+from torchapp import Param, method, tool, TorchApp
 from .modelsx import GambitModel
-from .dataloaders import create_dataloaders, seqbank_dataloader
-from .embedding import get_key
 from .gtdbtk import read_tophits, read_tigrfam, read_pfam
-from .embeddings.esm import ESMEmbedding
+from .embedding import get_key
 
-RANKS = ["phylum", "class", "order", "family", "genus", "species"]
+
 
 console = Console()
 
-class Gambit(ta.TorchApp):
-    """
-    Geometric Analysis of MicroBIal Taxonomies
-    """
-    def dataloaders(
+def read_memmap(path, count, dtype:str="float16") -> np.memmap:
+    file_size = os.path.getsize(path)
+    dtype_size = np.dtype(dtype).itemsize
+    num_elements = file_size // dtype_size
+    embedding_size = num_elements // count
+    shape = (count, embedding_size)
+    return np.memmap(path, dtype=dtype, mode='r', shape=shape)
+
+
+def gene_id_from_accession(accession:str):
+    return accession.split("/")[-1]
+
+RANKS = ["phylum", "class", "order", "family", "genus", "species"]
+
+esm_layers = 6
+DOMAIN = "bac120"
+# DOMAIN = "ar53"
+
+
+@dataclass(kw_only=True)
+class GambitDataset(Dataset):
+    accessions: list[str]
+    array:np.memmap|np.ndarray
+    gene_id_dict: dict[str, int]
+    accession_to_array_index:dict[str,int]|None=None
+
+    def __len__(self):
+        return len(self.accessions)
+
+    def __getitem__(self, idx):
+        accession = self.accessions[idx]
+        array_index = self.accession_to_array_index[accession] if self.accession_to_array_index else idx
+        embedding = torch.as_tensor(np.array(self.array[array_index,:], copy=False), dtype=torch.float16)
+        # embedding = torch.tensor(self.array[array_index,:], dtype=torch.float16)
+        gene_id = gene_id_from_accession(accession)
+        return embedding, self.gene_id_dict[gene_id]
+
+
+# @dataclass(kw_only=True)
+# class GambitTrainingDataset(GambitDataset):
+#     seqtree: SeqTree
+
+#     def __getitem__(self, idx):
+#         result = super()[idx]
+#         accession = self.accessions[idx]
+#         return *result, self.seqtree[accession].node_id
+
+
+@dataclass(kw_only=True)
+class GambitTrainingDataset(Dataset):
+    accessions: list[str]
+    seqtree: SeqTree
+    array:np.memmap|np.ndarray
+    gene_id_dict: dict[str, int]
+    accession_to_array_index:dict[str,int]|None=None
+
+    def __post_init__(self):
+        print('GambitTrainingDataset', hex(self.array.ctypes.data))
+
+    def __len__(self):
+        return len(self.accessions)
+
+    def __getitem__(self, idx):
+        accession = self.accessions[idx]
+        array_index = self.accession_to_array_index[accession] if self.accession_to_array_index else idx
+        embedding = torch.zeros( (320,), dtype=torch.float16) # hack
+        # embedding = torch.as_tensor(np.array(self.array[array_index,:], copy=False), dtype=torch.float16)
+        gene_id = gene_id_from_accession(accession)
+        return embedding, self.gene_id_dict[gene_id], self.seqtree[self.accessions[0]].node_id # hack
+        # return embedding, self.gene_id_dict[gene_id], self.seqtree[accession].node_id
+
+
+@dataclass
+class GambitDataModule(L.LightningDataModule):
+    seqtree: SeqTree
+    # seqbank: SeqBank
+    array:np.memmap|np.ndarray
+    accession_to_array_index:dict[str,int]
+    gene_id_dict: dict[str,int]
+    max_items: int = 0
+    batch_size: int = 16
+    num_workers: int = 0
+    validation_partition:int = 0
+
+    def __init__(
         self,
-        batch_size:int = ta.Param(default=32, help="The batch size."),
-        seqbank:Path = None,
-        seqtree:Path = None,
-        validation_partition:int=0,
-        max_items:int=None,
-        esm_layers:int=6,
-    ) -> DataLoaders:
-        """
-        Creates a FastAI DataLoaders object which Gambit uses in training and prediction.
+        seqtree: SeqTree,
+        array:np.memmap|np.ndarray,
+        accession_to_array_index:dict[str,int],
+        # seqbank: SeqBank,
+        gene_id_dict: dict[str,int],
+        max_items: int = 0,
+        batch_size: int = 16,
+        num_workers: int = 0,
+        validation_partition:int = 0,
+    ):
+        super().__init__()
+        self.array = array
+        self.accession_to_array_index = accession_to_array_index
+        self.seqtree = seqtree
+        self.gene_id_dict = gene_id_dict
+        self.max_items = max_items
+        self.batch_size = batch_size
+        self.validation_partition = validation_partition
+        self.num_workers = num_workers or os.cpu_count()
 
-        Args:
-            inputs (Path): The input file.
-            batch_size (int, optional): The number of elements to use in a batch for training and prediction. Defaults to 32.
+    def setup(self, stage=None):
+        # make assignments here (val/train/test split)
+        # called on every process in DDP
+        self.training = []
+        self.validation = []
 
-        Returns:
-            DataLoaders: The DataLoaders object.
-        """
-        assert seqbank is not None
-        print(f"Loading seqbank {seqbank}")
-        seqbank = SeqBank(seqbank)
+        for accession, details in self.seqtree.items():
+            partition = details.partition
+            dataset = self.validation if partition == self.validation_partition else self.training
+            dataset.append( accession )
+
+            if self.max_items and len(self.training) >= self.max_items and len(self.validation) > 0:
+                break
+
+        self.train_dataset = self.create_dataset(self.training)
+        self.val_dataset = self.create_dataset(self.validation)
+
+    def create_dataset(self, accessions:list[str]) -> GambitTrainingDataset:
+        return GambitTrainingDataset(
+            accessions=accessions, 
+            seqtree=self.seqtree, 
+            array=self.array,
+            accession_to_array_index=self.accession_to_array_index,
+            gene_id_dict=self.gene_id_dict,
+        )
+    
+    def train_dataloader(self):
+        print('train dataloader', self.num_workers)
+        return DataLoader(self.train_dataset, batch_size=self.batch_size, num_workers=self.num_workers, shuffle=True)
+
+    def val_dataloader(self):
+        print('val_dataloader', self.num_workers)
+        return DataLoader(self.val_dataset, batch_size=self.batch_size, num_workers=self.num_workers, shuffle=False)
+
+
+class Gambit(TorchApp):
+    @method
+    def setup(
+        self,
+        memmap:str=None,
+        memmap_index:str=None,
+        seqtree:str=None,
+        in_memory:bool=False,
+        prune:str="",
+    ) -> None:
+        if not seqtree:
+            raise ValueError("seqtree is required")
+        if not memmap:
+            raise ValueError("memmap is required")
+        if not memmap_index:
+            raise ValueError("memmap_index is required")        
 
         print(f"Loading seqtree {seqtree}")
-        seqtree = SeqTree.load(seqtree)
+        self.seqtree = SeqTree.load(seqtree)
 
-        self.classification_tree = seqtree.classification_tree
+        # prune classification tree if requested
+        if prune:
+            assert prune in RANKS[:-1]
+            prune_rank_index = RANKS.index(prune) + 1
+            for key, node_detail in self.seqtree.items():
+                node = self.seqtree.node(key)
+                ancestors = node.ancestors
+                if len(ancestors) > prune_rank_index:
+                    new_node = node.ancestors[prune_rank_index]
+                    self.seqtree[key] = SeqDetail(new_node, node_detail.partition)
+            
+            # remove children of genus nodes
+            for node in self.seqtree.classification_tree.pre_order_iter():
+                self.readonly = False
+                if len(node.ancestors) == prune_rank_index:
+                    node.children = []
+
+            self.seqtree.set_indexes()
+            RANKS = RANKS[:-1]# hack
+
+
+        # assert seqbank is not None
+        # print(f"Loading seqbank {seqbank}")
+        # self.seqbank = SeqBank(seqbank)
+        print(f"Loading memmap")
+        dtype = "float16"
+        self.accession_to_array_index = dict()
+        with open(memmap_index) as f:
+            for i, accession in enumerate(f):
+                self.accession_to_array_index[accession.strip()] = i
+        count = len(self.accession_to_array_index)
+        self.array = read_memmap(memmap, count)
+
+        # If there's enough memory, then read into RAM
+        if in_memory:
+            self.array = np.array(self.array)
+
+
+
+
+        self.classification_tree = self.seqtree.classification_tree
         assert self.classification_tree is not None
 
-        # HACK This should be general
-        embedding = ESMEmbedding(layers=esm_layers)
+        # Get list of gene families
+        family_ids = set()
+        for accession in self.seqtree:
+            gene_id = accession.split("/")[-1]
+            family_ids.add(gene_id)
 
-        dataloaders = create_dataloaders(
-            seqbank=seqbank,
-            seqtree=seqtree,
-            batch_size=batch_size,
-            validation_partition=validation_partition,
-            max_items=max_items,
-            embedding=embedding,
-        )
-        self.gene_id_dict = dataloaders.gene_id_dict
-        return dataloaders
+        self.gene_id_dict = {family_id:index for index, family_id in enumerate(sorted(family_ids))}
 
+    @method
     def model(
         self,
-        features:int=5120,
-        intermediate_layers:int=0, 
+        features:int=1024,
+        intermediate_layers:int=2,
         growth_factor:float=2.0,
-        family_embedding_size:int=64,
+        family_embedding_size:int=128,
     ) -> nn.Module:
-        """
-        Creates a deep learning model for the Gambit to use.
-
-        Returns:
-            nn.Module: The created model.
-        """
         return GambitModel(
             classification_tree=self.classification_tree,
             features=features,
@@ -92,38 +260,67 @@ class Gambit(ta.TorchApp):
             family_embedding_size=family_embedding_size,
             gene_family_count=len(self.gene_id_dict),
         )
-
-    def loss_func(self):
+    
+    @tool
+    def input_count(self) -> int:
+        return 2
+            
+    @method
+    def loss_function(self):
+        def dummy_loss(prediction, target):
+            return prediction.mean() * 0.0
+        return dummy_loss # hack
         return HierarchicalSoftmaxLoss(root=self.classification_tree)
     
-    def metrics(self):
-        return [
-            GreedyAccuracy(root=self.classification_tree, max_depth=i+1, name=rank)
-            for i, rank in enumerate(RANKS)
-        ]
-
-    def monitor(self):
-        return RANKS[-2] # i.e. genus
-
-    def inference_dataloader(
+    @method    
+    def metrics(self) -> list[tuple[str,Metric]]:
+        return [] # hack
+        rank_accuracy = RankAccuracyTorchMetric(
+            root=self.classification_tree, 
+            ranks={1+i:rank for i, rank in enumerate(RANKS)},
+        )
+                
+        return [('rank_accuracy', rank_accuracy)]
+        # return [
+        #     (rank, GreedyAccuracyTorchMetric(root=self.classification_tree, max_depth=i+1, name=rank))
+        #     for i, rank in enumerate(RANKS)
+        # ]
+    
+    @method    
+    def data(
         self,
-        learner,
+        max_items:int=0,
+        num_workers:int=0,
+    ) -> Iterable|L.LightningDataModule:
+        return GambitDataModule(
+            # seqbank=self.seqbank,
+            array=self.array,
+            accession_to_array_index=self.accession_to_array_index,
+            seqtree=self.seqtree,
+            gene_id_dict=self.gene_id_dict,
+            max_items=max_items,
+            num_workers=num_workers,
+        )
+    
+    @method
+    def prediction_dataloader(
+        self,
+        module,
         gtdbtk_output:Path=None,
-        seqbank:Path=None,
+        embeddings:Path=None,
         fasta:Path=None,
         tophits:Path=None,
         tigrfam:Path=None,
         pfam:Path=None,
         batch_size:int = 64,
-        **kwargs,
-    ):
-        assert seqbank is not None
-        print(f"Loading seqbank {seqbank}")
-        seqbank = SeqBank(seqbank)
-
+        num_workers: int = 0,
+    ) -> Iterable:
         # Get Embedding model from dataloaders
-        embedding = learner.dls.embedding
+        embedding = module.embedding        
         assert embedding is not None
+
+        self.classification_tree = module.classification_tree
+        assert self.classification_tree is not None
 
         # If gtdbtk output directory is given, then fetch the top hits from tigrfam
         def get_subfile(directory:Path, pattern:str) -> Path|None:
@@ -148,13 +345,17 @@ class Gambit(ta.TorchApp):
             gene_family_dict = read_pfam(pfam)
 
         # Find genes where we need to create the embeddings
-        accessions = {f"{gene_id}/{family_id}" for gene_id, family_id in gene_family_dict.items()}
-        accessions_to_do = accessions - set(seqbank.get_accessions())
-        genes_to_do = {accession.split("/")[0] for accession in accessions_to_do}
+        accessions = list({f"{gene_id}/{family_id}" for gene_id, family_id in gene_family_dict.items()})
+        genes_to_do = [accession.split("/")[0] for accession in accessions]
+        count = len(accessions)
 
-        # Generate embeddings if necessary
-        print(f"Generating embeddings for {len(genes_to_do)} protein sequences")
-        if genes_to_do:
+        dtype = "float16"
+        if Path(embeddings).exists():
+            array = read_memmap(embeddings, count, dtype=dtype)
+        else:
+            # Generate embeddings if necessary
+            array = None
+            print(f"Generating embeddings for {len(accessions)} protein sequences")
             assert fasta is not None
             fasta = Path(fasta)
             assert fasta.exists()
@@ -163,32 +364,41 @@ class Gambit(ta.TorchApp):
                     family_id = gene_family_dict[record.id]
                     print(record.id, family_id)
                     vector = embedding(record.seq)
-                    if vector is not None and not torch.isnan(vector).any():
-                        accession = f"{record.id}/{family_id}"
-                        seqbank.add(
-                            seq=vector.cpu().detach().clone().numpy().tobytes(),
-                            accession=accession,
-                        )
+                    if vector is not None and not torch.isnan(vector).any():                        
+                        # Initialise array if necessary
+                        if array is None:
+                            size = len(vector)
+                            shape = (count,size)
+                            array = np.memmap(embeddings, dtype=dtype, mode='w+', shape=shape)
 
-        self.items = list(accessions - set(seqbank.get_accessions()))
-        self.dataloader = seqbank_dataloader(seqbank=seqbank, items=self.items, batch_size=batch_size)
-        self.classification_tree = learner.dls.classification_tree
-        return self.dataloader
+                        # save vector to array
+                        index = genes_to_do.index(record.id)
+                        array[index] = vector.cpu().detach().clone().numpy()
 
+        dataset = GambitDataset(
+            accessions=accessions, 
+            array=self.array,
+            gene_id_dict=self.gene_id_dict,
+        )
+
+        self.items = accessions
+
+        return DataLoader(dataset, batch_size=batch_size, num_workers=self.num_workers, shuffle=False)
+
+    @method
     def output_results(
-        self,
-        results,
-        output_csv: Path = ta.Param(default=None, help="A path to output the results as a CSV."),
-        output_gene_csv: Path = ta.Param(default=None, help="A path to output the results for individual genes as a CSV."),
-        output_tips_csv: Path = ta.Param(default=None, help="A path to output the results as a CSV which only stores the probabilities at the tips."),
-        image: Path = ta.Param(default=None, help="A path to output the result as an image."),
+        self, 
+        results, 
+        output_csv: Path = Param(default=None, help="A path to output the results as a CSV."),
+        output_gene_csv: Path = Param(default=None, help="A path to output the results for individual genes as a CSV."),
+        output_tips_csv: Path = Param(default=None, help="A path to output the results as a CSV which only stores the probabilities at the tips."),
+        image: Path = Param(default=None, help="A path to output the result as an image."),
         image_threshold:float = 0.005,
-        prediction_threshold:float = ta.Param(default=0.0, help="The threshold value for making hierarchical predictions."),
+        prediction_threshold:float = Param(default=0.0, help="The threshold value for making hierarchical predictions."),
         seqtree:Path = None,
         output_correct:Path=None,
         **kwargs,
     ):
-        
         assert self.classification_tree # This should be saved from the learner
 
         # Sum the scores which is equivalent of multiplying the probabilities assuming that they are independent
@@ -275,8 +485,3 @@ class Gambit(ta.TorchApp):
 
         return results_df
 
-    def node_to_str(self, node:SoftmaxNode) -> str:
-        """ 
-        Converts the node to a string
-        """
-        return str(node).split(",")[-1].strip()
